updated documentation and added error checking to RedHelixClientReport.
