--- redhx-build-all/redhx-app-server-db/nbactions.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<actions>
-    <action>
-        <actionName>run</actionName>
-        <packagings>
-            <packaging>jar</packaging>
-        </packagings>
-        <goals>
-            <goal>process-classes</goal>
-            <goal>org.codehaus.mojo:exec-maven-plugin:1.2.1:exec</goal>
-        </goals>
-        <properties>
-            <exec.args>-Dparam_protocol="http" -Dparam_hostname="localhost" -Dparam_port="9080" -classpath %classpath org.redhelix.server.main.RedHelixServerDb</exec.args>
-            <exec.executable>java</exec.executable>
-        </properties>
-    </action>
-    <action>
-            <actionName>debug</actionName>
-            <packagings>
-                <packaging>jar</packaging>
-            </packagings>
-            <goals>
-                <goal>process-classes</goal>
-                <goal>org.codehaus.mojo:exec-maven-plugin:1.2.1:exec</goal>
-            </goals>
-            <properties>
-                <exec.args>-Xdebug -Xrunjdwp:transport=dt_socket,server=n,address=${jpda.address} -Dparam_protocol="http" -Dparam_hostname="localhost" -Dparam_port="9080" -classpath %classpath org.redhelix.server.main.RedHelixServerDb</exec.args>
-                <exec.executable>java</exec.executable>
-                <jpda.listen>true</jpda.listen>
-            </properties>
-        </action>
-    <action>
-            <actionName>profile</actionName>
-            <packagings>
-                <packaging>jar</packaging>
-            </packagings>
-            <goals>
-                <goal>process-classes</goal>
-                <goal>org.codehaus.mojo:exec-maven-plugin:1.2.1:exec</goal>
-            </goals>
-            <properties>
-                <exec.args>-Dparam_protocol="http" -Dparam_hostname="localhost" -Dparam_port="9080" -classpath %classpath org.redhelix.server.main.RedHelixServerDb</exec.args>
-                <exec.executable>java</exec.executable>
-            </properties>
-        </action>
-</actions>
