+++ redhx-build-all/redhx-server-db/src/main/webapp/META-INF/context.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Context antiJARLocking="true" path="/redhx-server-db"/>
